src/data_preperation/requirements.txt
requests
pyyaml
lxml
beautifulsoup4
PyGithub
GitPython
